后端/app/requirements.txt

flask
flask-cors
openai
whisper
tqdm
python-dotenv
transformers
Pillow
torch
numpy
scikit-learn
pytest
requests